requirements.txt
google-adk==1.4.1
google-genai>=1.17.0
python-dotenv>=1.1.0